content/pkg/deps/python/requirements.txt
aws==0.2.5
aws-lambda-powertools==1.18.1
aws-cdk.core==1.116.0
bcrypt==3.2.0
boto==2.49.0
cffi==1.14.5
cfn-flip==1.2.3
cryptography==3.4.7
fabric==2.6.0
importlib-metadata==4.5.0
invoke==1.5.0
paramiko==2.7.2
pathlib2==2.3.5
prettytable==2.1.0
pycparser==2.20
pydantic==1.8.2
PyYAML==5.4.1
PyNaCl==1.4.0
requests==2.25.1
six==1.16.0
typing-extensions==3.10.0.0
wcwidth==0.2.5
zipp==3.4.1
